+++ lib/views/view_categories/view_categories.dart
@@ -6,6 +6,7 @@ import 'package:money/storage/data/data.dart';
 import 'package:money/models/money_objects/categories/category.dart';
 import 'package:money/models/money_objects/currencies/currency.dart';
 import 'package:money/models/money_objects/transactions/transaction.dart';
+import 'package:money/views/view_categories/merge_categories.dart';
 import 'package:money/widgets/center_message.dart';
 import 'package:money/widgets/three_part_label.dart';
 import 'package:money/widgets/chart.dart';
@@ -24,6 +25,18 @@ class ViewCategories extends ViewForMoneyObjects {
 class ViewCategoriesState extends ViewForMoneyObjectsState {
   final List<Widget> pivots = <Widget>[];
   final List<bool> _selectedPivot = <bool>[false, false, false, false, false, true];
+
+  ViewCategoriesState() {
+    onMergeToItem = (final BuildContext context, final MoneyObject selectedObject) {
+      // let the user pick another Category and move the transactions of the current selected Category to the destination
+      showDialog(
+        context: context,
+        builder: (context) {
+          return MergeCategoriesTransactionsDialog(categoryToMove: selectedObject as Category);
+        },
+      );
+    };
+  }
 
   @override
   void initState() {