+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [version 1.8.9] 2024-05-19
+
+### Added
+
+- Feature for Merging|Appending categories
+
 ## [version 1.8.8] 2024-05-18
 
 ### Added